requirements.txt
openai==1.79.0
elevenlabs==0.2.27
streamlit==1.33.0
python-dotenv==1.0.1